# Performance Optimization: Improved TileShape Configuration for Large Llama Shapes (#3790)